refactoring day 23 for yielding int_code
removing translate functions to utils, as they are generic
springdroid part 2 answer said part 1
zork is the other consumer of translate, change it to utils
