+++ aoc2019/day25/zork.py
@@ -1,6 +1,5 @@
-import re
+import utils
 from aoc2019.int_code import IntCode
-import aoc2019.springscript as springscript
 
 NEWLINE = '\n'
 EXIT = 'exit'
@@ -14,12 +13,12 @@ COMPUTER = IntCode('input.txt')
 PROGRAM = COMPUTER.load_program()
 RESULT = PROGRAM.send(None)
 while isinstance(RESULT, list):
-    input_string = input(springscript.translate_to_chars(RESULT) + NEWLINE)
+    input_string = input(utils.translate_to_chars(RESULT) + NEWLINE)
     if input_string == EXIT:
         break
     input_string += NEWLINE
     for char in input_string:
         RESULT = PROGRAM.send(ord(char))
 FILE = open('answer_pt_1.txt', 'w')
-FILE.write(springscript.translate_to_chars(RESULT))
+FILE.write(utils.translate_to_chars(RESULT))
 FILE.close()
